requirements.txt
agora_token_builder==1.0.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==3.1.0
Django==5.1.7
django-cors-headers==4.7.0
django-fcm==0.1.1
django-geoposition==0.3.0
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
fcm-django==2.3.1
firebase_admin==7.1.0
geographiclib==2.0
geopy==2.4.1
gevent==25.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
locust-cloud==1.24.2
msgpack==1.1.0
packaging==25.0
pillow==11.1.0
pip-autoremove==0.10.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
redis==5.2.1
requests==2.32.4
rsa==4.9.1
service-identity==24.2.0
setuptools==76.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
swapper==1.4.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.event==5.1
zope.interface==7.2
whitenoise==6.11.0
